Inject proxies in rendered sections
